backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.1.4
langchain-community==0.0.16
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2
pypdf==4.0.1
PyMuPDF==1.23.21
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
qdrant-client==1.7.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redis==5.0.1
hiredis==2.3.2
python-dotenv==1.0.0
wandb==0.16.2
numpy==1.26.3
pandas==2.1.4
tqdm==4.66.1
python-json-logger==2.0.7
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==23.12.1
flake8==7.0.0
# RAG Enhancement Dependencies
ragas==0.1.0
rank-bm25==0.2.2
supabase==2.3.4
openai==1.10.0
# anthropic==0.18.0  # Optional - only if using Claude models
tiktoken==0.5.2
sse-starlette==2.0.0
nltk==3.8.1
scikit-learn==1.4.0
